Move deployment status-checking logic to library package
(cherry picked from commit 540f40871d643705fd491b2881028e9e1ee718fd)
